requirements.txt
Pillow==6.2.1
numpy==1.17.3
